Intent: Corregir importaciones redundantes Operators involved: [Coherence] Affected invariants: [#4] Key changes: - Normalizar import de threading en pruebas de caché de step Expected risks/dissonances: Ninguno, cobertura mínima Metrics: No aplica Equivalence map: tnfr.program -> tnfr.execution (pruebas)
