+++ src/tnfr/program.py
@@ -1,50 +1,33 @@
-"""TNFR programming language.
-
-Public exports are declared in ``__all__`` for explicit star imports.
-"""
+"""Backward compatible façade for the TNFR programming helpers."""
 
 from __future__ import annotations
-from typing import Any, Optional, Union
-from dataclasses import dataclass
-from collections import deque
-from collections.abc import Callable, Iterable, Sequence
-from functools import lru_cache
-from enum import Enum, auto
 
-import networkx as nx  # networkx is used at runtime
-
-from .token_parser import validate_token
-
-from .constants import get_param
-from .grammar import apply_glyph_with_grammar
-from .constants_glyphs import GLYPHS_CANONICAL_SET
-from .types import Glyph
-from .collections_utils import (
-    ensure_collection,
-    MAX_MATERIALIZE_DEFAULT,
-    is_non_string_sequence,
-    flatten_structure,
+from .execution import (
+    AdvanceFn,
+    HANDLERS,
+    _apply_glyph_to_targets,
+    _handle_glyph,
+    _handle_target,
+    _handle_thol,
+    _handle_wait,
+    _record_trace,
+    basic_canonical_example,
+    block,
+    get_step_fn,
+    play,
+    seq,
+    target,
+    wait,
 )
-from .glyph_history import ensure_history
-
-# Basic types
-Node = Any
-AdvanceFn = Callable[[Any], None]  # normalmente dynamics.step
-
-HandlerFn = Callable[
-    [nx.Graph, Any, Optional[list[Node]], deque, AdvanceFn],
-    Optional[list[Node]],
-]
-
-
-@lru_cache(maxsize=1)
-def get_step_fn() -> AdvanceFn:
-    """Return the dynamics ``step`` function, caching the import."""
-
-    from .dynamics import step as step_impl
-
-    return step_impl
-
+from .flatten import (
+    THOLEvaluator,
+    _flatten,
+    _flatten_glyph,
+    _flatten_target,
+    _flatten_wait,
+)
+from .token_parser import validate_token
+from .tokens import OpTag, TARGET, THOL, WAIT, THOL_SENTINEL, Token
 
 __all__ = (
     "WAIT",
@@ -63,401 +46,3 @@ __all__ = (
     "play",
     "basic_canonical_example",
 )
-
-# ---------------------
-# DSL constructs
-# ---------------------
-
-
-@dataclass(slots=True)
-class WAIT:
-    """Wait a number of steps without applying glyphs.
-
-    Attributes:
-        steps: number of steps the system advances.
-    """
-
-    steps: int = 1
-
-
-@dataclass(slots=True)
-class TARGET:
-    """Select the subset of nodes for subsequent glyphs.
-
-    Attributes:
-        nodes: iterable of nodes; ``None`` for all nodes.
-    """
-
-    nodes: Optional[Iterable[Node]] = None  # None = all nodes
-
-
-@dataclass(slots=True)
-class THOL:
-    """THOL block that opens self-organisation.
-
-    Attributes:
-        body: sequence of tokens to execute within the block.
-        repeat: how many times to repeat ``body``.
-        force_close: ``Glyph.SHA`` or ``Glyph.NUL`` to force closure.
-    """
-
-    body: Sequence[Any]
-    repeat: int = 1  # number of times to repeat the body
-    force_close: Optional[Glyph] = (
-        None  # None → automatic closure; SHA or NUL to force
-    )
-
-
-Token = Union[Glyph, WAIT, TARGET, THOL, str]
-
-# Sentinel used internally by ``_flatten`` to mark the end of a ``THOL`` block
-THOL_SENTINEL = object()
-
-
-class OpTag(Enum):
-    TARGET = auto()
-    WAIT = auto()
-    GLYPH = auto()
-    THOL = auto()
-
-
-# ---------------------
-# Internal utilities
-# ---------------------
-
-
-def _window(G) -> int:
-    return int(get_param(G, "GLYPH_HYSTERESIS_WINDOW"))
-
-
-# ---------------------
-# Execution core
-# ---------------------
-
-
-def _apply_glyph_to_targets(
-    G, g: Glyph | str, nodes: Optional[Iterable[Node]] = None
-):
-    """Apply ``g`` to ``nodes`` (or all nodes) respecting the grammar.
-
-    ``nodes`` may be any iterable of nodes, including a ``NodeView`` or other
-    iterables. To avoid unnecessary materialisation, iteration happens over the
-    iterable as-is; callers must materialise it if indexing is needed.
-    """
-    nodes_iter = G.nodes() if nodes is None else nodes
-    w = _window(G)
-    apply_glyph_with_grammar(G, nodes_iter, g, w)
-
-
-def _advance(G, step_fn: AdvanceFn):
-    step_fn(G)
-
-
-# ---------------------
-# Sequence compilation → list of atomic operations
-# ---------------------
-
-
-def _push_thol_frame(
-    frames: list[dict[str, Any]],
-    item: THOL,
-    *,
-    max_materialize: int | None,
-) -> None:
-    """Validate ``item`` and append a frame for its evaluation."""
-
-    repeats = int(item.repeat)
-    if repeats < 1:
-        raise ValueError("repeat must be ≥1")
-    if item.force_close is not None and not isinstance(item.force_close, Glyph):
-        raise ValueError("force_close must be a Glyph")
-    closing = (
-        item.force_close
-        if isinstance(item.force_close, Glyph)
-        and item.force_close in {Glyph.SHA, Glyph.NUL}
-        else None
-    )
-    seq0 = ensure_collection(
-        item.body,
-        max_materialize=max_materialize,
-        error_msg=f"THOL body exceeds max_materialize={max_materialize}",
-    )
-    frames.append(
-        {
-            "seq": seq0,
-            "index": 0,
-            "remaining": repeats,
-            "closing": closing,
-        }
-    )
-
-
-class THOLEvaluator:
-    """Generador que expande un bloque :class:`THOL`."""
-
-    def __init__(
-        self,
-        item: THOL,
-        *,
-        max_materialize: int | None = MAX_MATERIALIZE_DEFAULT,
-    ) -> None:
-        self._frames: list[dict[str, Any]] = []
-        _push_thol_frame(
-            self._frames, item, max_materialize=max_materialize
-        )
-        self._max_materialize = max_materialize
-        self._started = False
-
-    def __iter__(self) -> "THOLEvaluator":
-        return self
-
-    def __next__(self):
-        if not self._started:
-            self._started = True
-            return THOL_SENTINEL
-        while self._frames:
-            frame = self._frames[-1]
-            seq = frame["seq"]
-            idx = frame["index"]
-            if idx < len(seq):
-                token = seq[idx]
-                frame["index"] = idx + 1
-                if isinstance(token, THOL):
-                    _push_thol_frame(
-                        self._frames,
-                        token,
-                        max_materialize=self._max_materialize,
-                    )
-                    return THOL_SENTINEL
-                return token
-            else:
-                cl = frame["closing"]
-                frame["remaining"] -= 1
-                if frame["remaining"] > 0:
-                    frame["index"] = 0
-                else:
-                    self._frames.pop()
-                if cl is not None:
-                    return cl
-        raise StopIteration
-
-
-def _flatten_target(
-    item: TARGET,
-    ops: list[tuple[OpTag, Any]],
-) -> None:
-    ops.append((OpTag.TARGET, item))
-
-
-def _flatten_wait(
-    item: WAIT,
-    ops: list[tuple[OpTag, Any]],
-) -> None:
-    steps = max(1, int(getattr(item, "steps", 1)))
-    ops.append((OpTag.WAIT, steps))
-
-
-def _flatten_glyph(
-    item: Glyph | str,
-    ops: list[tuple[OpTag, Any]],
-) -> None:
-    g = item.value if isinstance(item, Glyph) else str(item)
-    if g not in GLYPHS_CANONICAL_SET:
-        raise ValueError(f"Non-canonical glyph: {g}")
-    ops.append((OpTag.GLYPH, g))
-
-
-_TOKEN_DISPATCH: dict[type, Callable[[Any, list[tuple[OpTag, Any]]], None]] = {
-    TARGET: _flatten_target,
-    WAIT: _flatten_wait,
-    Glyph: _flatten_glyph,
-    str: _flatten_glyph,
-}
-
-
-def _flatten(
-    seq: Sequence[Token],
-    *,
-    max_materialize: int | None = MAX_MATERIALIZE_DEFAULT,
-) -> list[tuple[OpTag, Any]]:
-    """Return list of operations ``(op, payload)``.
-    ``op`` ∈ :class:`OpTag`.
-
-    Parameters
-    ----------
-    seq:
-        Sequence of tokens to flatten.
-    max_materialize:
-        Maximum number of items to materialize from ``seq`` or ``THOL`` bodies
-        when they are not already collections. Defaults to
-        :data:`MAX_MATERIALIZE_DEFAULT`; ``None`` disables the limit.
-    """
-
-    ops: list[tuple[OpTag, Any]] = []
-    sequence = ensure_collection(seq, max_materialize=max_materialize)
-
-    def _expand(item: Any):
-        if isinstance(item, THOL):
-            return THOLEvaluator(item, max_materialize=max_materialize)
-        return None
-
-    for item in flatten_structure(sequence, expand=_expand):
-        if item is THOL_SENTINEL:
-            ops.append((OpTag.THOL, Glyph.THOL.value))
-            continue
-        handler = _TOKEN_DISPATCH.get(type(item))
-        if handler is None:
-            raise TypeError(f"Unsupported token: {item!r}")
-        handler(item, ops)
-    return ops
-
-
-# ---------------------
-# Handlers for atomic tokens
-# ---------------------
-
-
-def _record_trace(trace: deque, G, op: OpTag, **data) -> None:
-    trace.append({"t": float(G.graph.get("_t", 0.0)), "op": op.name, **data})
-
-
-def _advance_and_record(
-    G,
-    trace: deque,
-    label: OpTag,
-    step_fn: AdvanceFn,
-    *,
-    times: int = 1,
-    **data,
-) -> None:
-    for _ in range(times):
-        _advance(G, step_fn)
-    _record_trace(trace, G, label, **data)
-
-
-def _handle_target(
-    G, payload: TARGET, _curr_target, trace: deque, _step_fn: AdvanceFn
-):
-    """Handle a ``TARGET`` token and return the active node set.
-
-    Notes
-    -----
-    The node source is materialized with ``max_materialize=None`` so there is
-    no limit on the number of nodes retrieved.
-
-    Returns
-    -------
-    Collection[Node]
-        Collection of nodes to be used for subsequent operations.
-    """
-    nodes_src = G.nodes() if payload.nodes is None else payload.nodes
-    nodes = ensure_collection(nodes_src, max_materialize=None)
-    curr_target = nodes if is_non_string_sequence(nodes) else tuple(nodes)
-    _record_trace(trace, G, OpTag.TARGET, n=len(curr_target))
-    return curr_target
-
-
-def _handle_wait(
-    G, steps: int, curr_target, trace: deque, step_fn: AdvanceFn
-):
-    _advance_and_record(G, trace, OpTag.WAIT, step_fn, times=steps, k=steps)
-    return curr_target
-
-
-def _handle_glyph(
-    G,
-    g: str,
-    curr_target,
-    trace: deque,
-    step_fn: AdvanceFn,
-    label: OpTag = OpTag.GLYPH,
-):
-    _apply_glyph_to_targets(G, g, curr_target)
-    _advance_and_record(G, trace, label, step_fn, g=g)
-    return curr_target
-
-
-def _handle_thol(
-    G, g, curr_target, trace: deque, step_fn: AdvanceFn
-):
-    return _handle_glyph(
-        G, g or Glyph.THOL.value, curr_target, trace, step_fn, label=OpTag.THOL
-    )
-
-
-HANDLERS: dict[OpTag, HandlerFn] = {
-    OpTag.TARGET: _handle_target,
-    OpTag.WAIT: _handle_wait,
-    OpTag.GLYPH: _handle_glyph,
-    OpTag.THOL: _handle_thol,
-}
-
-
-# ---------------------
-# Public API
-# ---------------------
-
-
-def play(
-    G, sequence: Sequence[Token], step_fn: Optional[AdvanceFn] = None
-) -> None:
-    """Execute a canonical sequence on graph ``G``.
-
-    Rules:
-      - Use ``TARGET(nodes=...)`` to change the subset of application.
-      - ``WAIT(k)`` advances ``k`` steps with the current selector
-        (no forced glyph).
-      - ``THOL([...], repeat=r, force_close=…)`` opens a self-organising block,
-        repeats the body and optionally forces closure with SHA/NUL.
-      - Glyphs are applied via ``enforce_canonical_grammar``.
-    """
-    step_fn = step_fn or get_step_fn()
-
-    curr_target: Optional[list[Node]] = None
-
-    # Traza de programa en history
-    history = ensure_history(G)
-    maxlen = int(get_param(G, "PROGRAM_TRACE_MAXLEN"))
-    trace = history.get("program_trace")
-    if not isinstance(trace, deque) or trace.maxlen != maxlen:
-        trace = deque(trace or [], maxlen=maxlen)
-        history["program_trace"] = trace
-
-    for op, payload in _flatten(sequence):
-        handler: HandlerFn | None = HANDLERS.get(op)
-        if handler is None:
-            raise ValueError(f"Unknown operation: {op}")
-        curr_target = handler(G, payload, curr_target, trace, step_fn)
-
-
-# ---------------------
-# Helpers to build sequences easily
-# ---------------------
-
-
-def seq(*tokens: Token) -> list[Token]:
-    return list(tokens)
-
-
-def block(
-    *tokens: Token, repeat: int = 1, close: Optional[Glyph] = None
-) -> THOL:
-    return THOL(body=list(tokens), repeat=repeat, force_close=close)
-
-
-def target(nodes: Optional[Iterable[Node]] = None) -> TARGET:
-    return TARGET(nodes=nodes)
-
-
-def wait(steps: int = 1) -> WAIT:
-    return WAIT(steps=max(1, int(steps)))
-
-
-def basic_canonical_example() -> list[Token]:
-    """Reference canonical sequence.
-
-    SHA → AL → RA → ZHIR → NUL → THOL
-    """
-    return seq(
-        Glyph.SHA, Glyph.AL, Glyph.RA, Glyph.ZHIR, Glyph.NUL, Glyph.THOL
-    )
